Refactor: 구성 파일을 수정한다.
