+++ src/main/java/org/prgms/springbootbasic/common/Console.java
@@ -5,7 +5,8 @@ import org.prgms.springbootbasic.domain.VoucherType;
 import org.springframework.stereotype.Component;
 
 import java.text.MessageFormat;
-import java.util.*;
+import java.util.Collection;
+import java.util.Scanner;
 
 import static org.prgms.springbootbasic.common.CommonConstant.INPUT_FIXED_AMOUNT_VOUCHER;
 import static org.prgms.springbootbasic.common.CommonConstant.INPUT_PERCENT_DISCOUNT_VOUCHER;
