+++ src/main/java/org/prgms/springbootbasic/common/file/CsvFileTemplate.java
@@ -2,6 +2,7 @@ package org.prgms.springbootbasic.common.file;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 import java.io.*;
@@ -11,6 +12,7 @@ import java.util.List;
 import java.util.function.Function;
 
 @Component
+@Profile("test")
 public class CsvFileTemplate {
     private static final Logger log = LoggerFactory.getLogger(CsvFileTemplate.class);
