+++ src/main/java/org/prgms/springbootbasic/common/file/CustomerCsvFileManager.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 import static org.prgms.springbootbasic.common.CommonConstant.CSV_PATTERN;
 
 @Component
-@Profile({"dev", "prod"})
+@Profile({"test"})
 @Slf4j
 public class CustomerCsvFileManager {
     private static final String BLACK_PATH = "./src/main/resources/customer_blacklist.csv";
