+++ src/main/java/org/prgms/springbootbasic/repository/customer/CustomerDatabaseRepository.java
@@ -9,7 +9,6 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import java.nio.ByteBuffer;
 import java.time.LocalDateTime;
 import java.util.*;
