+++ src/main/resources/application-dev.yml
@@ -1,10 +1,1 @@
-spring:
-  datasource:
-    url: jdbc:mysql://localhost:3306/test
-    username: test
-    password: 1234
-    driver-class-name: com.mysql.cj.jdbc.Driver
 
-basic:
-  file:
-    path: ./src/test/resources/voucher.csv
