--- src/main/resources/application-prod.yml
@@ -1,3 +1,0 @@
-basic:
-  file:
-    path: ./src/main/resources/voucher.csv
