+++ src/main/resources/application.yml
@@ -1,3 +1,8 @@
 spring:
   profiles:
     active: prod
+  datasource:
+    url: jdbc:mysql://localhost:3306/test
+    username: test
+    password: 1234
+    driver-class-name: com.mysql.cj.jdbc.Driver
